Also move rulebased test utils to dedicated endpoint
