+++ feature-libs/product-configurator/rulebased/cpq/rest/cpq-configurator-utils.spec.ts
@@ -1,11 +1,11 @@
 import { TestBed } from '@angular/core/testing';
 import { Configurator } from '@spartacus/product-configurator/rulebased';
-import { ConfiguratorTestUtils } from '../../shared/testing';
+import { ConfiguratorTestUtils } from '@spartacus/product-configurator/rulebased/testing';
 import {
   attributeCheckbox,
   attributeRadioButton,
   productConfiguration,
-} from '../../shared/testing/configurator-test-data';
+} from '../../testing/configurator-test-data';
 import { CpqConfiguratorEndpointService } from './cpq-configurator-endpoint.service';
 import { CpqConfiguratorUtils } from './cpq-configurator-utils';
