+++ feature-libs/product-configurator/rulebased/index.ts
@@ -2,4 +2,3 @@ export * from './components/index';
 export * from './core/index';
 export * from './occ/index';
 export * from './rulebased-configurator.module';
-export * from './shared/testing/index';
