+++ feature-libs/product-configurator/rulebased/occ/variant/converters/occ-configurator-variant-normalizer.spec.ts
@@ -5,10 +5,10 @@ import {
   OccConfig,
   TranslationService,
 } from '@spartacus/core';
+import { ConfiguratorTestUtils } from '@spartacus/product-configurator/rulebased/testing';
 import { Observable, of } from 'rxjs';
 import { OccConfigurator } from '../variant-configurator-occ.models';
 import { Configurator } from './../../../core/model/configurator.model';
-import { ConfiguratorTestUtils } from './../../../shared/testing/configurator-test-utils';
 import { OccConfiguratorVariantNormalizer } from './occ-configurator-variant-normalizer';
 
 const configId = '192826';
