+++ feature-libs/product-configurator/rulebased/testing/ng-package.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "../../../../node_modules/ng-packagr/ng-package.schema.json",
+  "lib": {
+    "entryFile": "./public_api.ts",
+    "umdModuleIds": {
+      "@spartacus/core": "core"
+    }
+  }
+}
