+++ feature-libs/product-configurator/rulebased/testing/public_api.ts
@@ -1,0 +1,1 @@
+export * from './index';
