+++ projects/storefrontapp/tsconfig.app.prod.json
@@ -66,11 +66,14 @@
       "@spartacus/product-configurator/rulebased": [
         "dist/product-configurator/rulebased"
       ],
+      "@spartacus/product-configurator/rulebased/cpq": [
+        "dist/product-configurator/rulebased/cpq"
+      ],
       "@spartacus/product-configurator/rulebased/root": [
         "dist/product-configurator/rulebased/root"
       ],
-      "@spartacus/product-configurator/rulebased/cpq": [
-        "dist/product-configurator/rulebased/cpq"
+      "@spartacus/product-configurator/rulebased/testing": [
+        "dist/product-configurator/rulebased/testing"
       ],
       "@spartacus/product-configurator/textfield": [
         "dist/product-configurator/textfield"
