+++ tsconfig.compodoc.json
@@ -101,12 +101,15 @@
       "@spartacus/product-configurator/rulebased": [
         "feature-libs/product-configurator/rulebased/public_api"
       ],
+      "@spartacus/product-configurator/rulebased/cpq": [
+        "feature-libs/product-configurator/rulebased/cpq/public_api"
+      ],
       "@spartacus/product-configurator/rulebased/root": [
         "feature-libs/product-configurator/rulebased/root/public_api"
       ],
-      "@spartacus/product-configurator/rulebased/cpq": [
-        "feature-libs/product-configurator/rulebased/cpq/public_api"
-      ],      
+      "@spartacus/product-configurator/rulebased/testing": [
+        "feature-libs/product-configurator/rulebased/testing/public_api"
+      ],
       "@spartacus/product-configurator/textfield": [
         "feature-libs/product-configurator/textfield/public_api"
       ],
